backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
requests
python-dotenv